pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.135.1",
    "groq>=1.1.1",
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.3.0",
    "pyspellchecker>=0.9.0",
    "pytesseract>=0.3.13",
    "sentence-transformers>=5.3.0",
]
